Use createFromInterface for more reliable DateTime mapping
