+++ tests/Transformer/DateTimeInterfaceToMutableTransformerTest.php
@@ -4,21 +4,21 @@ declare(strict_types=1);
 
 namespace AutoMapper\Tests\Transformer;
 
-use AutoMapper\Transformer\DateTimeMutableToImmutableTransformer;
+use AutoMapper\Transformer\DateTimeInterfaceToMutableTransformer;
 use PHPUnit\Framework\TestCase;
 
-class DateTimeMutableToImmutableTransformerTest extends TestCase
+class DateTimeInterfaceToMutableTransformerTest extends TestCase
 {
     use EvalTransformerTrait;
 
     public function testDateTimeImmutableTransformer(): void
     {
-        $transformer = new DateTimeMutableToImmutableTransformer();
+        $transformer = new DateTimeInterfaceToMutableTransformer();
 
-        $date = new \DateTime();
+        $date = new \DateTimeImmutable();
         $output = $this->evalTransformer($transformer, $date);
 
-        self::assertInstanceOf(\DateTimeImmutable::class, $output);
+        self::assertInstanceOf(\DateTime::class, $output);
         self::assertSame($date->format(\DateTime::RFC3339), $output->format(\DateTime::RFC3339));
     }
 }
